Use sendmail to email multiple people unique links
The IR team I work on has a problem where we're constantly dealing with
silly people putting passwords in JIRA or ServiceNow. Usually this
happens because there's a shared account with a rotating password, or
maybe a client has given a certain team FTP creds (no comment on the use
of FTP.)

It's a bit painful to use `go-flashpaper` for this case because you have
to make a new secret for each team member who needs the file, so this
way you can just put in a list of recipients.

The one caveat of my approach is that in order to keep things
lightweight I use the local MTA (`sendmail`) to send mail. In most
environments that should be fine - I also tested this by spinning up a
fresh EC2 t2.micro w/ Amazon Linux and `sendmail`ing to my personal
address - but there's definitely going to be cases where you deploy this on a
box that has outbound firewall rules preventing mail from being
delivered.

At the very least, I can confirm that a random EC2 public DNS entry has
a good enough sender reputation to deliver to my personal Gmail inbox,
which should suffice for most ragtag orgs.